Refactor dependencies
Previously, the dependencies were all locked to a specific version in
package.json. This makes it hard to update the dependencies since it
requires changing each line to change each dependency. Using '^' in
package.json allows for running 'yarn upgrade' to install all bugfix and
backwards compatible upgrades for all the packages. Using '--pure-lockfile'
when installing the packages allows for the lockfile to be the
controlling file for which versions of each package are officially being
used and which are used in reproducible builds.